Added IF decimation for less CPU usage and enhanced filters
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@162 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ Radio/PortAudio/WaveDuplex.cs
@@ -11,7 +11,7 @@ namespace SDRSharp.Radio.PortAudio
         private readonly AudioBufferAvailableDelegate _bufferAvailable;
         private readonly PaStreamCallbackDelegate _paCallback = PaStreamCallback;
 
-        public WaveDuplex(int deviceIndex, int sampleRate, int framesPerBuffer, AudioBufferAvailableDelegate bufferNeededDelegate)
+        public WaveDuplex(int deviceIndex, double sampleRate, int framesPerBuffer, AudioBufferAvailableDelegate bufferNeededDelegate)
         {
             _bufferAvailable = bufferNeededDelegate;
